requirements.txt
phidata
groq
python-dotenv
duckduckgo-search
Nest-asynciore
requests